[cmake] Always build the libLLVM shared library
Summary:
We discourage non-developers from using the BUILD_SHARED_LIBS option
and recommend using LLVM_BUILD_LLVM_DYLIB=ON for building LLVM as
a shared library.  We should make it as easy as possible for
users to build with our recommended configuration, so to help with that
this patch makes building libLLVM.so the default and removes the
LLVM_BUILD_LLVM_DYLIB option altogether.

Removing this option will help simplify the build system and also
help downstream users of LLVM, because they won't need to add
special logic to their own build systems to detect whether or not libLLVM
was built and installed.

Reviewers: beanz, smeenai, compnerd, phosek

Subscribers: mgorny, mehdi_amini, dexonsmith, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D70026