Add LLVM_ABI macros to all public classes and functions (NFC)
This patch annotates every class and function defined in the LLVM
headers with the LLVM_ABI macro.  This macro is left undefined, so
this entire patch should be a no-op.  In the future, we will define
the LLVM_ABI with the appropriate visibility attributes, but first
we will commit the macro addition to avoid having to do large reverts
if the macro definitions we add later cause problems.