[workflows] Fix lldb-tests and libclc-tests
This was broken by d25022bb689b9bf48a24c0ae6c29c1d3c2f32823, which
caused the workflow to pass an empty string to ninja as the target.
The 'all' target is probably not the right target for these tests,
but this is what the behavior was before
d25022bb689b9bf48a24c0ae6c29c1d3c2f32823.