[CMake] Add a cache file for generating code coverage statistics
This is a simple 2-stage cache file, which generates a code coverage
report for the llvm-project code base.  By default this will anlayze
the top-level directories: llvm, clang, lld, and clang-tools-extra.

This can be customized by using the BOOTSTRAP_LLVM_ENABLE_PROJECTS
option to specify the projects being built and the
BOOTSTRAP_LLVM_CODE_COVERAGE_TARGETS option which will produce
code coverage for the binaries associated with the given targets.